stop storing delegates as separate state entry
